[edited drill]

+++ c# course/classmethod_Drill3/classmethod_Drill3/math2.cs
@@ -8,7 +8,7 @@ namespace classmethod_Drill3
 {
     public class math2
     {
-            public float Subt(float Number2)
+            public double Subt(double Number2)
             {
                 return Number2 - .5;
             }
